move bookmark item to separate component
