+++ src/components/Verses/BookmarkedVersesList.tsx
@@ -2,7 +2,7 @@
 /* eslint-disable jsx-a11y/no-static-element-interactions */
 /* eslint-disable jsx-a11y/click-events-have-key-events */
 /* eslint-disable jsx-a11y/anchor-is-valid */
-import React, { useContext, useMemo } from 'react';
+import React, { useMemo } from 'react';
 
 import useTranslation from 'next-translate/useTranslation';
 import { shallowEqual, useSelector, useDispatch } from 'react-redux';
@@ -10,33 +10,25 @@ import useSWR from 'swr';
 
 import styles from './BookmarkedVersesList.module.scss';
 import BookmarkedVersesListSkeleton from './BookmarkedVesesListSkeleton';
+import BookmarkPill from './BookmarkPill';
 
 import Link from '@/dls/Link/Link';
 import { ToastStatus, useToast } from '@/dls/Toast/Toast';
-import CloseIcon from '@/icons/close.svg';
 import { selectBookmarks, toggleVerseBookmark } from '@/redux/slices/QuranReader/bookmarks';
 import { selectQuranReaderStyles } from '@/redux/slices/QuranReader/styles';
 import { getMushafId } from '@/utils/api';
 import { deleteBookmarkById, privateFetcher } from '@/utils/auth/api';
 import { makeBookmarksUrl } from '@/utils/auth/apiPaths';
 import { isLoggedIn } from '@/utils/auth/login';
-import { getChapterData } from '@/utils/chapter';
 import { logButtonClick } from '@/utils/eventLogger';
-import { toLocalizedVerseKey } from '@/utils/locale';
-import { getVerseNavigationUrlByVerseKey } from '@/utils/navigation';
-import {
-  getChapterNumberFromKey,
-  getVerseAndChapterNumbersFromKey,
-  makeVerseKey,
-} from '@/utils/verse';
-import DataContext from 'src/contexts/DataContext';
+import { getVerseAndChapterNumbersFromKey, makeVerseKey } from '@/utils/verse';
 import Bookmark from 'types/Bookmark';
 
 const BOOKMARKS_API_LIMIT = 10; // The number of bookmarks to fetch from the api
 
 const BookmarkedVersesList = () => {
-  const { t, lang } = useTranslation('home');
-  const chaptersData = useContext(DataContext);
+  const { t } = useTranslation('home');
+
   const quranReaderStyles = useSelector(selectQuranReaderStyles, shallowEqual);
   const dispatch = useDispatch();
 
@@ -104,10 +96,6 @@ const BookmarkedVersesList = () => {
     }
   };
 
-  const onLinkClicked = () => {
-    logButtonClick('bookmarked_verses_list_link');
-  };
-
   const onViewAllBookmarksClicked = () => {
     logButtonClick('view_all_bookmarks');
   };
@@ -129,34 +117,9 @@ const BookmarkedVersesList = () => {
       {bookmarkedVersesKeys.length > 0 ? (
         <div className={styles.bookmarksContainer}>
           <div className={styles.verseLinksContainer}>
-            {bookmarkedVersesKeys?.map((verseKey) => {
-              const chapterNumber = getChapterNumberFromKey(verseKey);
-              const chapterData = getChapterData(chaptersData, chapterNumber.toString());
-              const bookmarkText = `${chapterData.transliteratedName} ${toLocalizedVerseKey(
-                verseKey,
-                lang,
-              )}`;
-              return (
-                <div key={verseKey} className={styles.bookmarkItem}>
-                  <Link
-                    href={getVerseNavigationUrlByVerseKey(verseKey)}
-                    onClick={onLinkClicked}
-                    className={styles.linkButtonContainer}
-                  >
-                    {bookmarkText}
-                  </Link>
-                  <button
-                    onClick={onBookmarkDeleted(verseKey)}
-                    type="button"
-                    className={styles.closeIconContainer}
-                  >
-                    <span>
-                      <CloseIcon />
-                    </span>
-                  </button>
-                </div>
-              );
-            })}
+            {bookmarkedVersesKeys?.map((verseKey) => (
+              <BookmarkPill key={verseKey} verseKey={verseKey} onDeleted={onBookmarkDeleted} />
+            ))}
             {hasReachedBookmarksLimit && (
               <Link
                 href="/collections/all"
